backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pymongo==4.10.1
pydantic==2.10.6
python-multipart==0.0.20
python-dotenv==1.0.1
emergentintegrations
bcrypt==4.2.1
Pillow==10.4.0